Bump version for deploy to PyPi (#25)
I released to PyPi the version 2.0.0 off a branch. I changed the way we install PyGithub after that upload(deploy) to PyPi, so now that we added that change and merged it to master we need a new version. 

TL;DR - bump version because we now install PyGithub from PyPi not Github.com